+++ services/preview/reservations/mgmt/2019-04-01/reservations/CHANGELOG.md
@@ -1,2 +1,15 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. CalculatePriceResponseProperties.GrandTotal
+1. CalculatePriceResponseProperties.NetTotal
+1. CalculatePriceResponseProperties.TaxTotal
+1. PropertiesType.Archived
+1. PropertiesType.Capabilities
+1. PurchaseRequestProperties.InstanceFlexibility
+1. ScopeProperties.Reason